0.1.20

+++ core/src/linuxX64Main/kotlin/pw/binom/charset/AbstractIconv.kt
@@ -51,16 +51,16 @@ abstract class AbstractIconv(fromCharset: String, toCharset: String) : Closeable
 
                     outputPointer.ptr.reinterpret(),
                     outputAvail.ptr
-            )
+            ).toInt()
             val readed = beforeIn - inputAvail.value.toInt()
             val writed = beforeOut - outputAvail.value.toInt()
             input.position += readed
             output.position += writed
             return when {
-                r != 0uL && errno == E2BIG -> CharsetTransformResult.OUTPUT_OVER
-                r != 0uL && errno == EILSEQ -> CharsetTransformResult.MALFORMED
-                r != 0uL && errno == EINVAL -> CharsetTransformResult.INPUT_OVER
-                r == 0uL && errno == 0 -> CharsetTransformResult.SUCCESS
+                r != 0 && errno == E2BIG -> CharsetTransformResult.OUTPUT_OVER
+                r != 0 && errno == EILSEQ -> CharsetTransformResult.MALFORMED
+                r != 0 && errno == EINVAL -> CharsetTransformResult.INPUT_OVER
+                r == 0 && errno == 0 -> CharsetTransformResult.SUCCESS
                 else -> throw IllegalStateException()
             }
         }
